% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recall_at_k.R
\name{recall_at_k}
\alias{recall_at_k}
\title{Recall@K}
\usage{
recall_at_k(scores, outs, grid)
}
\arguments{
\item{scores}{Scores of (nominal) outlyingness. A higher score here implies an observation is more likely to be an outlier.}

\item{outs}{Vector of outlier indices.}

\item{grid}{Grid of K values over which the recall is computed. Must be between 0 and 1.}
}
\value{
Recall values at the points of the provided grid.
}
\description{
Function computing the recall based on the top K\% scores of outlyingness.
}
\examples{
\donttest{
dt <- as.data.frame(sample(c(1:2), 100, replace = TRUE, prob = c(0.5, 0.5)))
dt <- cbind(dt, sample(c(1:3), 100, replace = TRUE, prob = c(0.5, 0.3, 0.2)))
dt[, 1] <- as.factor(dt[, 1])
dt[, 2] <- as.factor(dt[, 2])
colnames(dt) <- c('V1', 'V2')
sono_out <- sono(data = dt,
probs = list(c(0.5, 0.5), c(1/3, 1/3, 1/3)),
alpha = 0.01,
r = 2,
MAXLEN = 0,
frequent = FALSE)
# Suppose observations 1 up to 5 are outliers
recall_at_k(scores = sono_out[[2]][, 2],
outs = c(1:5),
grid = c(1, 2.5, seq(5, 50, by = 5))/100)
}
}
