% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc_auc.R
\name{roc_auc}
\alias{roc_auc}
\title{ROC AUC function}
\usage{
roc_auc(scores, outs, grid)
}
\arguments{
\item{scores}{Scores of (nominal) outlyingness. A higher score here implies an observation is more likely to be an outlier.}

\item{outs}{Vector of outlier indices.}

\item{grid}{Grid of Top K values over which the ROC AUC is computed. Must be between 0 and 1.}
}
\value{
ROC AUC  at the points of the provided grid.
}
\description{
Function computing the ROC AUC given a vector with scores of outlyingness. The computation for this is based on
\insertCite{hanley_meaning_1982;textual}{SONO}.
}
\examples{
dt <- as.data.frame(sample(c(1:2), 100, replace = TRUE, prob = c(0.5, 0.5)))
dt <- cbind(dt, sample(c(1:3), 100, replace = TRUE, prob = c(0.5, 0.3, 0.2)))
dt[, 1] <- as.factor(dt[, 1])
dt[, 2] <- as.factor(dt[, 2])
colnames(dt) <- c('V1', 'V2')
sono_out <- sono(data = dt,
probs = list(c(0.5, 0.5), c(1/3, 1/3, 1/3)),
alpha = 0.01,
r = 2,
MAXLEN = 0,
frequent = FALSE)
# Suppose observations 1 up to 5 are outliers
roc_auc(scores = sono_out[[2]][, 2], outs = c(1:5),
grid = c(1, 2.5, seq(5, 50, by = 5))/100)

}
\references{
{
\insertRef{hanley_meaning_1982}{SONO}
}
}
